+++ src/pages/connectYanoljaPage/IntroPage/IntroPage.style.ts
