+++ src/pages/connectYanoljaPage/successPage/SuccessPage.style.ts
