+++ src/pages/connectYanoljaPage/verificationPage/VerificationPage.style.ts
