+++ src/pages/connectYanoljaPage/verificationPage/components/submitButton/SubmitButton.style.ts
