+++ src/pages/connectYanoljaPage/verificationPage/components/termsAgreementSection/TermsAgreementSection.style.ts
