+++ src/pages/connectYanoljaPage/verificationPage/components/verificationSection/VerificationSection.style.ts
