+++ src/routes/router.tsx
@@ -1,5 +1,9 @@
+import IntroPage from "@pages/connectYanoljaPage/IntroPage/IntroPage.tsx";
+import SuccessPage from "@pages/connectYanoljaPage/successPage/SuccessPage.tsx";
+import ConnectYanoljaAccount from "@pages/connectYanoljaPage/verificationPage/VerificationPage.tsx";
 import { Suspense } from "react";
 import { createBrowserRouter } from "react-router-dom";
+
 import App from "../App";
 import { PATH } from "../constants/path";
 
@@ -9,19 +13,17 @@ import Search from "../pages/searchPage";
 import SignUp from "../pages/signUpPage/SignUp";
 import MyPage from "../pages/myPage/MyPage";
 import SignIn from "../pages/signInPage/SignIn";
-import TransferWriting from "../pages/transferWritingPage/TransferWriting";
-import TransferSale from "../pages/transferSalePage";
 import RoomDetail from "@pages/roomDetailPage/RoomDetail";
 import TransferPurchase from "../pages/transferPurchasePage/TransferPurchase";
-import TransferWritingPrice from "@/pages/transferWritingPricePage/TransferWritingPrice";
-import PasswordReset from "@/pages/passwordResetPage/PasswordReset";
-import ConnectYanolja from "@pages/connectYanoljaPage/ConnectYanolja";
-import ConnectYanoljaAccount from "@pages/connectYanoljaAccountPage/ConnectYanoljaAccount";
-import ConnectYanoljaSuccess from "@pages/connectYanoljaSuccessPage/ConnectYanoljaSuccess";
-import Setting from "@/pages/myPage/setting/Setting";
-import ManageProfile from "@/pages/myPage/manage/manageProfile/ManageProfile";
+import TransferSale from "../pages/transferSalePage";
+import TransferWriting from "../pages/transferWritingPage/TransferWriting";
+
 import ManageAccount from "@/pages/myPage/manage/manageAccount/ManageAccount";
+import ManageProfile from "@/pages/myPage/manage/manageProfile/ManageProfile";
+import Setting from "@/pages/myPage/setting/Setting";
+import PasswordReset from "@/pages/passwordResetPage/PasswordReset";
 import PurchaseDetail from "@/pages/purchaseDetailPage/PurchaseDetail";
+import TransferWritingPrice from "@/pages/transferWritingPricePage/TransferWritingPrice";
 
 export const router = createBrowserRouter([
   {
@@ -108,15 +110,17 @@ export const router = createBrowserRouter([
       },
       {
         path: PATH.YANOLJA_ACCOUNT,
-        element: <ConnectYanolja />,
-      },
-      {
-        path: PATH.YANOLJA_ACCOUNT_VERIFY,
-        element: <ConnectYanoljaAccount />,
-      },
-      {
-        path: PATH.YANOLJA_ACCOUNT_VERIFY + "/success",
-        element: <ConnectYanoljaSuccess />,
+        element: <IntroPage />,
+        children: [
+          {
+            path: PATH.YANOLJA_ACCOUNT_VERIFY,
+            element: <ConnectYanoljaAccount />,
+          },
+          {
+            path: PATH.YANOLJA_ACCOUNT_VERIFY + "/success",
+            element: <SuccessPage />,
+          },
+        ],
       },
     ],
   },
